Remove custom Tooltip
