+++ frontend/src/components/user/SearchQuerySaveUserControl.tsx
@@ -3,13 +3,8 @@
 import { useTranslations } from "next-intl";
 import dynamic from "next/dynamic";
 
-// import SearchSavedQuery2 from "src/components/search/SearchSavedQuery2";
 const SearchSavedQuery = dynamic(
   () => import("src/components/search/SearchSavedQuery"),
-  { ssr: false },
-);
-const SearchSavedQuery2 = dynamic(
-  () => import("src/components/search/SearchSavedQuery2"),
   { ssr: false },
 );
 
@@ -17,10 +12,6 @@ export const SearchQuerySaveUserControl = () => {
   const t = useTranslations("Search.savedQuery");
   return (
     <>
-      <SearchSavedQuery2
-        copyText={t("copy.unauthenticated")}
-        helpText={t("help.unauthenticated")}
-      />
       <SearchSavedQuery
         copyText={t("copy.unauthenticated")}
         helpText={t("help.unauthenticated")}
